+++ lib/subvisual/math_utils.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+require "subvisual/math_utils/greatest_common_divisor"
+
+module Subvisual
+  module MathUtils
+    extend GreatestCommonDivisor
+  end
+end