+++ lib/subvisual/utils.rb
@@ -1,10 +1,7 @@
 # frozen_string_literal: true
 
-require "subvisual/utils/version"
-
 module Subvisual
   module Utils
-    class Error < StandardError; end
-    # Your code goes here...
+    VERSION = "0.1.0"
   end
 end